Move token method from front/helpers to common.erc20tokens
