+++ src/front/shared/helpers/web3.ts
@@ -2,7 +2,6 @@ import Web3 from 'web3'
 import config from 'app-config'
 import helpers from 'helpers'
 
-console.log('reset web3')
 let web3: IEtheriumProvider = new Web3(
   new Web3.providers.HttpProvider(
     (config.binance)
